Factor out linux-user/qemu.h
A quite large part of {linux,bsd}-user/qemu.h is shared, this patch
introduces qemu-user-common.h which factors it out from linux-user. This
shared part is also the bare minimum required to build a *-user-like
target, and, in particular, it will be useful for the libtcg targets.

At this point we don't factor out this code from bsd-user since
currently it's not in a stable shape, and verifying if these changes
affect its functionality is non-trivial. Once bsd-user will be back in a
working shape, we can proceed factoring its code out.