Factor out tcg/tcg.h
This patch factors out all the necessary data structures contained in
tcg.h into include/tcg-common.h so that they can be made available to
users of the library version of the TCG, libtcg.

The data structures have been modified in a way that allows them to be
easily prefixed in different ways depending on the situation. For
internal use, the prefix will typically be TCG_ for constants and TCG
for data structures, while it will be LIBTCG_/LibTCG for external use.

In addition to this, certain data types (namely TCGReg, TCGRegSet and
tcg_temp) have been forced to be the same over all the targets, so that
the data structures are architecture-independent, a vital requirement to
provide a uniform interface to libtcg users.

Finally tcg/tcg-opc.h has been moved to include/.